+++ tracing-core/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.7 (October 16, 2019)
+
+### Added
+
+- Added `dispatcher::set_default` API which returns a drop guard (#388)
+
 # 0.1.6 (September 12, 2019)
 
 ### Added